Continue with base price
This is almost the same, but there is a wrinkle in the form of the pesky call on super (which is pretty common in these kinds of subclass extension cases). When I move the subclass code to the delegate, I’ll need to call the parent case—but I can’t just call this._host._basePrice without getting into an endless recursion.

I have a couple of options here. One is to apply Extract Function (106) on the base calculation to allow me to separate the dispatch logic from price calculation. (The rest of the move is as before.)